project_1_delivery_delay_prediction/requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
